Wextra

+++ src/client/model/ClientModel.cpp
@@ -81,7 +81,7 @@ void ClientModel::removeObsoleteGameObjects()
     }
   }
 
-  for (int i = 0; i < destroyCandidates.size(); ++i)
+  for (size_t i = 0; i < destroyCandidates.size(); ++i)
   {
     logger << "removing " + std::to_string(destroyCandidates[i]);
     if (gameObjects[destroyCandidates[i]]->id == clientInfo.id)
